+++ package.json
@@ -11,7 +11,7 @@
   },
   "files": [
     "build/",
-    "abi/",
+    "src/abi/",
     "src/",
     "typechain/"
   ],
@@ -60,14 +60,13 @@
     "test:coverage": "c8 -c .nycrc mocha -r ts-node/register tests/**/*.test.ts",
     "test:wait": "ts-node tests/wait.ts",
     "test": "mocha -r ts-node/register tests/**/*.test.ts",
-    "build": "yarn types && tsc && ncp ./abi ./build/abi",
+    "build": "tsc",
     "lint": "gts lint",
     "lint:fix": "gts fix",
     "watch": "tsc --watch",
     "types:fetch": "cd scripts && ./update-abi.sh && cd ../",
-    "types": "typechain --target ethers-v6 --out-dir typechain abi/*.json",
+    "types": "typechain --target ethers-v6 --out-dir src/typechain src/abi/*.json",
     "clean": "gts clean",
-    "compile": "tsc",
     "docs": "typedoc --out docs --hideInPageTOC true --useCodeBlocks true --includeVersion true  src/index.ts && cd scripts && ./post-process-doc.sh && cd ../"
   },
   "publishConfig": {
