+++ src/abi/IERC1271.json
