+++ src/typechain/IERC1271.ts
