+++ src/typechain/factories/IERC1271__factory.ts
